requirements.txt
pyTelegramBotAPI == 4.7.1
PyYAML == 6.0